% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{all_cause_us}
\alias{all_cause_us}
\title{All Cause Deaths in the United States from 1999 to 2016}
\format{a data.frame with 18 rows and 5 columns.  Note that non-syntacticaly valid names are used for some of the
columns.  You will need to use backticks to refernce the columns by name.

\describe{
  \item{Year}{The calendar year}
  \item{Cause Name}{Cause of Death}
  \item{State}{Reporting State or Region}
  \item{Deaths}{Reported number of deaths}
  \item{Age-adjusted Death Rate}{No specifics on the method for this
  adjustment, but it is provided by NCHS}
}}
\usage{
all_cause_us
}
\description{
All Cause Deaths in the United States from 1999 to 2016
}
\seealso{
vignette("datasets", package = "drug.eg")
}
\keyword{datasets}
