% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{by_cause_states}
\alias{by_cause_states}
\title{Top 10 Causes of Deaths within the United States by State from 1999 to 2016}
\format{a data.frame with 9,180 rows and 5 columns.  Note that non-syntacticaly valid names are used for some of the
columns.  You will need to use backticks to refernce the columns by name.

\describe{
  \item{Year}{The calendar year}
  \item{Cause Name}{Cause of Death}
  \item{State}{Reporting State or Region}
  \item{Deaths}{Reported number of deaths}
  \item{Age-adjusted Death Rate}{No specifics on the method for this
  adjustment, but it is provided by NCHS}
}}
\usage{
by_cause_states
}
\description{
Top 10 Causes of Deaths within the United States by State from 1999 to 2016
}
\seealso{
vignette("datasets", package = "drug.eg")
}
\keyword{datasets}
